condense multiple fetcher files
