+++ frontend/src/fetchers/event.js
@@ -1,4 +1,7 @@
+import join from 'url-join';
 import { get, put } from './index';
+
+const eventsUrl = join('/', 'api', 'events');
 
 export const eventById = async (id) => {
   const res = await get(`/api/events/id/${id}`);
@@ -12,5 +15,11 @@ export const updateEvent = async (id, data) => {
 
 export const sessionsByEventId = async (id) => {
   const res = await get(`/api/session-reports/eventId/${String(id)}`);
+  return res.json();
+};
+
+export const getEventsByStatus = async (status, filters) => {
+  const url = join(eventsUrl, status, `?${filters}`);
+  const res = await get(url);
   return res.json();
 };
